+++ Sources/MpcNET/Commands/Database/ReadPictureCommand.cs
@@ -14,14 +14,14 @@ namespace MpcNET.Commands.Database
     public class ReadPictureCommand : IMpcCommand<MpdBinaryData>
     {
         private readonly string path;
-        private readonly int binaryOffset;
+        private readonly long binaryOffset;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ReadPictureCommand"/> class.
         /// </summary>
         /// <param name="path">The URI.</param>
         /// <param name="offset">Binary data offset if needed</param>
-        public ReadPictureCommand(string path, int offset = 0)
+        public ReadPictureCommand(string path, long offset = 0)
         {
             this.path = path;
             this.binaryOffset = offset;
@@ -47,8 +47,8 @@ namespace MpcNET.Commands.Database
             if (response.ResponseValues.Count == 0)
                 return null;
 
-            var totalSize = int.Parse(response.ResponseValues.Where(kvp => kvp.Key == "size").Select(kvp => kvp.Value).First()); 
-            var payloadSize = int.Parse(response.ResponseValues.Where(kvp => kvp.Key == "binary").Select(kvp => kvp.Value).First());
+            var totalSize = long.Parse(response.ResponseValues.Where(kvp => kvp.Key == "size").Select(kvp => kvp.Value).First()); 
+            var payloadSize = long.Parse(response.ResponseValues.Where(kvp => kvp.Key == "binary").Select(kvp => kvp.Value).First());
 
             return new MpdBinaryData(totalSize, payloadSize, response.BinaryData);
         }
